Move architecture dependent files into engine/arch
Bench: 4385715
