+++ engine/includes.hpp
@@ -17,8 +17,9 @@
 #include <unordered_set>
 #include <utility>
 
-#include "pzstl/vector.hpp"
+#include "arch/arch.hpp"
 #include "debug.hpp"
+#include "pzstl/vector.hpp"
 
 #ifndef VERSION
 #define VERSION "v7.0"
