+++ engine/nnue/network.cpp
@@ -1,5 +1,4 @@
 #include "network.hpp"
-#include "simd.hpp"
 
 #include "incbin.h"
